requirements.txt
Flask
tensorflow
numpy
opencv-python
pillow